+++ NEWS.md
@@ -6,6 +6,7 @@
 * 6 palettes have been added from the suffrager package. (#86)
 * 19 palettes have been added from the colorBlindness package. (#85)
 * 15 palettes have been added from the rtist package. (#84)
+* Some palette names have been changed to remove non-ascii characters.
 
 # paletteer 1.1.0